get connection using Name
